RF: Move all patool code into a dedicated code file
This shall make providing an alternative but API compatible
implementation easier to place and test alongside of this existing
one.
